+++ include/sealib/graph.h
@@ -17,7 +17,7 @@
 class Graph {
 private:
     Node* nodes;
-    size_t order;
+    unsigned int order;
 
 public:
     /**
@@ -25,14 +25,14 @@ public:
      * @param _nodes Array of nodes, to be initialized by the caller.
      * @param _order Order of the graph, equals the length of the nodes array.
      */
-    Graph(Node *_nodes, size_t _order);
+    Graph(Node *_nodes, unsigned int _order);
 
     /**
      * Getter for a specific node in the nodes array.
      * @param u index in the nodes array
      * @return Pointer to the Node in the nodes array
      */
-    Node* getNode(size_t u);
+    Node* getNode(unsigned int u);
 
     /**
     * Returns the index of the head of u in the nodes vector.
@@ -40,12 +40,12 @@ public:
     * @param k index in the adjacency vector of node u
     * @return index of the head vertex in the nodes vector
     */
-    size_t head(size_t u, size_t k);
+    unsigned int head(unsigned int u, unsigned int k);
 
     /**
     * Order is the number of nodes in the graph.
     * @return order of the graph as int
     */
-    size_t getOrder();
+    unsigned int getOrder();
 };
 #endif //SEA_GRAPH_Hs
